Fix to support Git source tags as well as branches
Currently, Luarocks does not support Git tags, only branches, because
it's relying on the --branch argument to clone when checking out a
repository. This means that when you pack a source rock that uses a Git
tag, Git resorts to using HEAD if the value set to source.tag in the
rockspec is a tag rather than an actual branch.

A better strategy is to use the `git checkout` command inside the cloned
git repository. This allows you to agnostically use either a tag or a
branch, which seems to be the desired functionality given the
documentation.