Adjust Git clone command according to Git version.
Git 1.7.10 and higher allows a branch or a tag to be passed as an
argument to `git clone` along with `--depth=1`, which lets us check out
the exact code needed to pack a rockspec without the entire repository's
history. This offers a significant performance improvement for large
repositories.

For Git 1.7.9 and earlier, Luarocks will clone the repository without
the branch argument in order to support both branches and tags. This is
possible in part because prior to Git 1.7.10, git will clone sufficient
repository history to check out any tag, even when the --depth option is
passed.

After cloning, Luarocks will then perform a `git checkout` prior
to packing a rockspec. This is slower but lets Luarocks work
consistently on more versions of Git.

Resolves #73
Resolves #78

This change was implemented by:

* Juan Ramírez (https://github.com/ichramm)
* Ignacio Burgueño (https://github.com/ignacio)
* Norman Clarke (https://github.com/norman)