Added nil check for tsfd in http_request()
If the opening of the cache timestamp file fails, then
tsfd would be nil.  write is then invoked on it yielding:

  /usr/local/share/lua/5.1/luarocks/fs/lua.lua:592: attempt to index local 'tsfd' (a nil value)

This changeset simply adds the check for nil.  It does no further error
reporting, so luarocks doesn't crash, but it also doesn't report that the
cache was not updated nor that anything out of the ordinary happened.