Default to a higher minimum OS X version.
In a month or two, OS X 10.10 is coming out, but the current maximum
MACOSX_DEPLOYMENT_TARGET is 10.5, which was released in 2007, which was
7 years ago. Most users are on a higher version of OS X now. Plus, 10.6
introduced ARC, which doesn't compile when MACOSX_DEPLOYMENT_TARGET <=
10.6, and which is an insanely useful feature. Currently, trying to
enable ARC in a project involves having to override "CC", which in
general is a bad idea since it should be left to the user to
configure. Removing MACOSX_DEPLOYMENT_TARGET from the "CC" variable is
another story, but for now this commit should be sufficient to make
enabling ARC much easier.