Fix installation of modules from files ending with "init.lua"
In builtin backend, when figuring out where to install a module and
checking for special "init.lua" logic, trigger it only when
the whole basename is "init.lua", not when it merely ends with it.

Fixes incorrect installation when build.modules contains pairs like
`["kainit"] = "kainit.lua"` - the module used to go to `kainit/kainit.lua`
instead of just `kainit.lua`.