Fix merge of rocks_provided [2].
This is now a parameter, because depending on the rockspec version the set of
default rocks_provided for a rockspec may change. This is to maintain
compatibility with rocks provided by LuaRocks versions in rockspec format 1.0.
When checking dependencies for a rockspec, we ensure we respect that to keep
backward compatibility. When we check dependencies for the manifest as a
whole, we can use the larger set of rocks_provided known by the current
LuaRocks version.