Breadth-first search of external dependencies.
Major refactoring of the external dependency search,
now including a breadth-first traversal of the external
dependency directories, meaning that if a library or
header is available at a subdirectory of one of the
configure paths, it will be found.

The recursion is limited to four levels deep to avoid
thrashing the disk too much in case a dependency is missing.
(In my system, scanning for a missing library takes 1.5s;
scanning for a library that is present one level deep
is essentially instantaneous).

Hopefully this means the end of the annoyances with
OPENSSL_DIR on various platforms!