configure: check presence of lua.h
Verification of the Lua include path happens at LuaRocks runtime,
but we also perform it here just so that the user gets an early failure
if they try to install LuaRocks with the Lua interpreter package
but not the "development files" that many Linux distros ship separately.

We also include a --disable-incdir-check flag for specialized scenarios: if
you do not wish to use "luarocks build", (e.g. when only deploying binary
packages) you do not need lua.h installed. This flag skips the check for lua.h
in "configure".