core.cfg: produce quoted values in tool variables for Windows
Alternative solution to the problem described in #968. We can't
use #968 because on Unix sometimes people set the value of those
variables so they include arguments (even LuaRocks does it
by default, setting `UNZIP="unzip -n"`).

Now that the variables are quoted, we stop auto-quoting them
in `luarocks.fs`. I expect this change to have to practical
impact on Unix, where paths with spaces are uncommon (those can
use explicit quotes in their values).