Fix and reorganize Lua version detection
Make it more robust when detecting the Lua version when working inside a
project, and hopefully make the detection logic easier to track in the
code. Reorganized code so that the `detected` table becomes internal
to the `init_config` operation and only the regular `cfg` global table
is used after that.

Includes regression tests.