fs: use Unix fork-free versions of is_file/is_dir in Linux and macOS only
The Unix fork-free version relies on non-standard behaviors. It works
on Linux, but it took a while for it to work properly in macOS, and it
turns out that you can't really properly detect a directory in FreeBSD
using io.open() because it can actually open a directory. To avoid filling
in platform-specific tricks in luarocks/fs/unix.lua, which was never the
goal, it's better to move the fork-free operations to Linux and macOS
specific backends, and keep other Unices using the 'test' command.