Ensure scripts add the prefix to the package path
This reverts the removal from e9215f139be7afa3af116a2039594362e4f1f2be
but keeps the addition of the custom package loader as that is
a more robust approach.

I'm not a super fan of this approach because in a cross-compiling
scenario it does leak the host package.path into the output of commands
such as `luarocks path` and building the `bin` wrappers, but at least
it is less broken than the 3.2.0 behavior, as reported by @fperrad in
issue #1070.