git: add support for shallow recommendations
In more recent versions of git, the .gitmodules file can
be configured to recommend if a submodule should be cloned as
shallow or not. Add support for this feature since it gives the
rock maintainer more control over which submodules should be shallow
cloned.

The rock maintainer may set the boolean `shallow` to true or false in
.gitmodules. If omitted it is treated as false.

Signed-off-by: Pedro Tammela <pctammela@gmail.com>