Add support for MSYS2 + Mingw-w64
Currently, LuaRocks supports:

  * (a) Lua interpreters built for MSYS2 (Lua interpreters depend on
    msys-2.0.dll).
    (the "msys" platform)

  * (b) Lua interpreters built by MinGW (Lua interpreters don't depend
    on msys-2.0.dll).
    (the "mingw" platform)

This change adds support for (c) Lua interpreters built as native
Windows application by MSYS2 + Mingw-w64 (Lua interpreters don't
depend on msys-2.0.dll). (the "msys2_mingw_w64" platform)

Here are differences between (a), (b) and (c):

  * (a) can't work without MSYS2 (msys-2.0.dll)
  * (b) can work without MSYS2
  * (c) can work without MSYS2 but is generally used with MSYS2
    because MSYS2 provides useful toolsets such as zip, unzip, wget
    and so on.

This change assumes that users use (c) with MSYS2. This change uses
zip, unzip, wget and so on provided by MSYS2.

MSYS2 has LuaRocks package:
https://github.com/msys2/MINGW-packages/tree/master/mingw-w64-lua-luarocks

It applies a patch to support (c). If this change is merged into
LuaRocks, MSYS2 doesn't need to have a patch for LuaRocks.

CI jobs are added for (c) but they ignore failures. Because the latest
LuaSocket has a problem with the latest Mingw-w64. We need
https://github.com/diegonehab/luasocket/pull/312 to fix this.