repos.delete_version: remove entry from manifest if rock is already missing
If rock_manifest could not be found, the entry in manifest is unusable:
without the list of files from rock_manifest, we can't scan the repository
to remove files. This means the entry in the repo manifest is a leftover from
an incomplete removal. Remove the entry from the repo manifest.