Set cfg.home_tree even for root user
Don't set cfg.home_tree to nil in cfg.lua when the user is root.
Instead, do explicit checks if the user is root in cmd.lua.

This removes some confusion, as cfg.home_tree is later set by the
default system config file, so it will show up when running 'luarocks
config'.