Support XDG_CONFIG_HOME
The loading of the config file has been changed to support the XDG
Base Directory specification. More info here:

  https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html

If in a UNIX environment, LuaRocks will try loading a config file in
XDG_CONFIG_HOME. If this fails, try loading one in the old location
~/.luarocks.

This method clashed a bit with the old code that separated the setting
of confdirs and the loading of the config files. As we now change
cfg.homeconfdir depending on where we find the config file I found it
necessary to remove the function set_confdirs() and move some pieces
into the config loading logic.