use mirrors when downloading rocks even if manifest succeeds
LuaRocks generally only checked whether to use a mirror in the first
operation, when it fetches the manifest. If the manifest fails to load, it
switches to the mirror and everything works from there. But if the manifest
fetches ok and the then actual rock download fails with a 504, it gives up,
instead of trying that in a mirror as well. Changing that to make it retry
every download on a mirror when the base URL matches one configured in
cfg.rocks_servers should make it much more resilient.

Fixes #1299.