loader.which: new option for searching package.path and cpath
Adds a new second argument, `where`, a string which indicates places
to search for the module.

If `where` contains "l", it will search using the LuaRocks loader; if it
contains "p", it will look in the filesystem using package.path and
package.cpath. You can use both at the same time.

If successful, it will return four values.

* If found using the LuaRocks loader, it will return:
   * filename of the module (eg. "/usr/local/lib/lua/5.1/socket/core.so"),
   * rock name
   * rock version
   * "l" to indicate the match comes from the loader.

* If found scanning package.path and package.cpath, it will return:
   * filename of the module (eg. "/usr/local/lib/lua/5.1/socket/core.so"),
   * "path" or "cpath"
   * nil
   * "p" to indicate the match comes from scanning package.path and cpath.

If unsuccessful, nothing is returned.