update copyright notice
I was involved with Kepler as a funded project from 2006-2009;
the project continued for a few years laters in a volunteer basis
without an actual legal entity. This outdated copyright notice
was maintained out of inertia, and I don't think anyone else
involved in Kepler back in the day would claim ownership of
LuaRocks beyond Kepler's original incarnation. To avoid any
misunderstandings moving forward and to give credit to the many
contributors over the years, I am updating this copyright
notice to reflect the actual existence of Kepler, having spoken
with original Kepler folks such as @carregal and @mascarenhas
about it. The "LuaRocks project authors" notice follows the
example of other projects such as V8; look at the Git history
for the detailed credits.