fix: install build_dependencies on the correct Lua version
When installing build dependencies, do so for the Lua version on which the
LuaRocks program is actually running, and not the one selected by the user via
config or `--lua-version`.

To do that, we temporarily flip the user-selected Lua version. It's a hacky
approach because we have to do this by flipping a bunch of global config
settings, and we may be missing some entries.

This definitely needs a better solution, but this is what can be done somewhat
easily in the current architecture. A full solution needs to address have
several complications (e.g. if you have a build dependency that needs to be
compiled, it will require C headers for another version of LuaRocks, and the
binary might be compiled with a different Lua version than the one the
developer has set up in their machine.)