feat(builtin): inherit inc and libdirs from external_dependencies
If a rockspec has external_dependencies but the module entry does
not specify incdirs or libdirs, then autoextract those values from
external_dependencies and apply it to the module entry.

Hopefully this will improve compatibility of existing rockspecs
that did not fully specify their incdirs and libdirs, such as
https://github.com/brimworks/lua-yajl/blob/078e48147e89d34b8224a07129675aa9b5820630/rockspecs/lua-yajl-2.0-1.rockspec

Fixes #1239.