Makefile: add work around for environments that default to multiple jobs
Fixes #1638

This Makefile only supports single-job execution and is known to have
race conditions that will fail if targets are run in parallel jobs. With
the proliferation of cores, more and more distributions and host
environments are setting default make flags to run jobs in parallel by
default. We can't do anything about this for old versions of GNU Make,
but for 4.3+ we can actually reset the flag ourselves to only run
a single job so targets run in serial order.