fix: when linking Lua explicitly (Windows), use dynamic libraries
The original code was inconsistent, since it demanded dynamic libraries
for Cygwin but static libraries for other Windows compilers. However,
that wouldn't work for the same reason as Cygwin:

From the LuaJIT makefile:
https://github.com/LuaJIT/LuaJIT/blob/ae4735f621d89d84758769b76432d2319dda9827/src/Makefile#L558-L563V

"Mixed mode [static + dynamic] is not supported on Windows.
And static mode doesn't work well.
C modules cannot be loaded, because they bind to lua51.dll."